Restructuring code to use Netmiko SCP
